Move tests out of naught spec
